review changes and changelog additions
