+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 All notable changes to this project will be documented in this file.
 ## [0.0.1-RC7]
 
-- Introduced utils classes for building SchemaAttributes from Class and validating a json payload
-  against the SchemaDetails of a specified schema key
 - Introduced a `leia-common` module to host all the common utils classes 
-- Removed the `SchemaValidatable` annotation and moved it to a generic `SchemaDefinition`
-- Few Bug fixes in SchemaValidationUtils 
+- Added annotation classes for the Qualifiers( PII, Encrypted, ShortLived) that can be added on the members of
+  the Schema class
+- Replaced the `SchemaValidatable` annotation and moved it to a generic `SchemaDefinition`
+- Introduced `SchemaBuilder`: For building the schema request against a class annotated with SchemaDefinition
+- Introduced `SchemaPayloadValidator`: For validating a schema json payload against a specified SchemaKey
+- Addressed issues in handling plain `Object` and Primitive Class types in `SchemaValidationUtils` 
 
 ## [0.0.1-RC6]
